requirements.txt
Flask==2.3.2
rembg==2.0.59  # Use the latest available version
Pillow==9.2.0